Pass `--no-divider` to `interpret-trailers`
This fixes the behavior for commits that include inline Git diffs in
Markdown fenced code blocks, like this:

```diff
diff --git a/src/main.rs b/src/main.rs
index dc9dfd4..1951719 100644
--- a/src/main.rs
+++ b/src/main.rs
@@ -232,2 +232,3 @@ fn trailers(message: String) -> err::Result<Vec<(String, String)>> {
             "--parse",
+            "--no-divider",
         ])
```

Previously, the `---` in the diff was erroneously treated as the divider
between a commit message and a patch. See `man git-interpret-trailers`
for more details.